requirements.txt
discord.py
python-dotenv
google-api-python-client
google-auth-oauthlib
ollama
pydantic
parsedatetime
